requirements.txt
boto3>=1.34.0,<2.0.0
requests>=2.32.0,<3.0.0
PyJWT[crypto]>=2.8.0,<3.0.0
pydantic>=2.7.0,<3.0.0
pytest>=8.0.0,<9.0.0
